+++ packages/classic/src/hashes-entry.ts
@@ -1,1 +1,0 @@
-export * from '@jscrypto/hashes';