+++ packages/kdfs/src/index.ts
@@ -1,5 +1,7 @@
+export * from './argon2.js';
 export * from './evpkdf.js';
 export * from './hkdf.js';
 export * from './hmac.js';
 export * from './pbkdf2.js';
 export * from './preset.js';
+export * from './scrypt.js';